Adore the beauty of kurtis
Adore the beauty in you. Kurtis is the perfect attire of choice for Indian women. It gives you more comfort and an elegant look. It can be worn by any age group.

LUCKNOWI KURTI- It is a classic piece that can be worn during the day with jeans, salwar & palazzos. It can be worn in a warmer climate. Chikankari is an eminent craft amongst other handloom specialities of India. It is also called shadow work. 

FLORAL KURTIS – Brighten your day up with this alluring floral Kurtis. Look attractive with our beautiful range of floral Kurtis. If you are looking for something elegant, comfortable, yet budget-friendly- floral Kurtis is the way to go. Floral one-piece gives a stunning look that gives an awesome look in the daytime or at a pool party.

PRINTED KURTI – Explore the latest women's taste for printed Kurtis. It will give you a stunning look. It is best suitable for the summer season as it is more comfortable for breathing and is an absorbent fabric like cotton, linen

Printed Kurtis are basic traditional ethnic. A long kurti or a short kurti both gives a good look. These fine quality ladies printed Kurtis are made by using very high and fine quality and skin-friendly fabric.

LYRA LEGGINGS - Shop for the latest & trendy lyra leggings. Lux Lyra Women's leggings and churidars are available in an ankle length. Lyra is a type of stretchy fabric, similar to elastic which is used to make tight-fitting garments.

WEDDING KURTIS - One of the most versatile dresses that women can wear at parties or at a festival or marriage. This traditional or Indo-western look is not just versatile but also elegant and gives a comfortable look. Women of all age groups wear it. Anarkali suits are evergreen and can be worn on all types of occasions. The perfect combination of elegance, grace and traditional quotient.

The demand for straight Kurtis is increasing massively day by day. The tall kurtis go best with Skinny jeans, Straight pants, and palazzo. It is a perfect office wear outfit.

NIGHT SUITS - Shop from the wide range of stylish & trendy night suits. Nightwear or sleepwear is designed to be worn while sleeping. The style of nightwear can change from season to season. woollen night dress is worn or made for winter days.